refactor(all): modularize into subpackages.
